+++ CHANGELOG.md
@@ -28,6 +28,7 @@ All notable changes to this project will be documented in this file.
 - DataGrid: Exception in AddDisplayControl #181
 - DataGrid: Support for custom type descriptors #200
 - PropertyGrid: Support BrowsableAttribute on enum items when shown as ComboBox or ListBox #133
+- PropertyGrid: Hidden error info when HeaderPlacement = Above #212
 
 ### Added
 - EditableAttribute